Move test for classes of instance variables to AppTest
Ruby 1.8.7 raises failure with message 'No tests were specified.' on unit tests without any method "test_*".
So I've decided to move tests for classes of fixtured objects to AppTest
(it is a main test, which is parent for all another tests). AppTest is used as "factory" for fixtured objects,
so it is suitable place for these tests